+++ server/data-service/src/main/java/edu/illinois/ncsa/incore/service/data/controllers/DatasetController.java
@@ -14,7 +14,6 @@ package edu.illinois.ncsa.incore.service.data.controllers;
 
 import edu.illinois.ncsa.incore.common.auth.IAuthorizer;
 import edu.illinois.ncsa.incore.common.auth.Privileges;
-import edu.illinois.ncsa.incore.common.config.Config;
 import edu.illinois.ncsa.incore.common.dao.ISpaceRepository;
 import edu.illinois.ncsa.incore.common.exceptions.IncoreHTTPException;
 import edu.illinois.ncsa.incore.common.models.Space;
@@ -78,8 +77,8 @@ import java.util.stream.Collectors;
 
 @Path("datasets")
 public class DatasetController {
-    private static final String DATA_REPO_FOLDER = Config.getConfigProperties().getProperty("data.repo.data.dir");
-    private static final String GEOSERVER_ENABLE = Config.getConfigProperties().getProperty("geoserver.enable");
+    private static final String DATA_REPO_FOLDER = System.getenv("DATA_REPO_DATA_DIR");
+    private static final String GEOSERVER_ENABLE = System.getenv("GEOSERVER_ENABLE");
     private static final String POST_PARAMETER_NAME = "name";
     private static final String POST_PARAMETER_FILE = "file";
     private static final String POST_PARAMETER_FILE_LINK = "link-file";